event manager dashboard reflect changes when you update the page
