+++ src/components/Cards/card.tsx
@@ -1,11 +1,14 @@
 import { useState } from "react";
 import Image from "next/image";
+import router from "next/router";
 import { RepairRequest } from "@prisma/client";
+import { SubmitHandler } from "react-hook-form";
 
 import AssigneeBadge from "@/components/Cards/assignee-badge";
 import StatusPill from "@/components/Cards/status-pill";
 import PrepopulatedRepairAttemptForm from "@/components/Forms/prepopulated-repair-request-form";
 import Modal from "@/components/Modal/index";
+import { GeneralRepairAttempt } from "@/types";
 
 export type CardProps = {
   title?: string;
@@ -25,6 +28,22 @@ export default function Card({ props }: { props: CardProps }) {
   }
 
   const [showModal, setShowModal] = useState(false);
+  const onSubmit: SubmitHandler<GeneralRepairAttempt> = async (data) => {
+    // console.log(JSON.stringify(data));
+    const response = await fetch(`/api/repair-request`, {
+      method: "PATCH",
+      headers: {
+        "Content-Type": "application/json"
+      },
+      body: JSON.stringify(data)
+    });
+    if (response.ok) {
+      router.reload();
+      setShowModal(false);
+    } else {
+      alert(`Error! ${response.statusText}`);
+    }
+  };
 
   return (
     <div
@@ -42,6 +61,7 @@ export default function Card({ props }: { props: CardProps }) {
           <div>
             <PrepopulatedRepairAttemptForm
               props={props.repairRequestProps}
+              onSubmit={onSubmit}
             ></PrepopulatedRepairAttemptForm>
           </div>
         </div>
